campus_safety_latest/model/requirements.txt
# YOLOv8 Exit Sign Detection Requirements
ultralytics>=8.0.0
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.7.0
pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0
numpy>=1.21.0
pyyaml>=6.0
tqdm>=4.64.0

# Optional: For advanced visualizations
plotly>=5.0.0
wandb>=0.13.0  # For experiment tracking

# For model conversion and optimization
onnx>=1.12.0
onnxruntime>=1.13.0
coremltools>=6.0  # For iOS deployment
tensorflow>=2.10.0  # For TensorFlow Lite conversion
